+++ doc_render_template.go
@@ -269,6 +269,24 @@ func docTemplateFuncs(parser *Parser, markHidden bool, format optionRenderFormat
 			}
 			return false
 		},
+
+		"hasNamedArgGroups": func(groups []docArgGroup) bool {
+			for _, group := range groups {
+				if group.Name != "" {
+					return true
+				}
+			}
+			return false
+		},
+
+		"hasNamedCommandGroups": func(groups []docCommandGroup) bool {
+			for _, group := range groups {
+				if group.Name != "" {
+					return true
+				}
+			}
+			return false
+		},
 	}
 }